fix refectory error.

+++ order-service/src/main/java/com/yiguan/jigsaw/order/services/command/OrderPaidCommand.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class OrderPaidCommand implements OrderCommand {
-  private Integer oid;
+  private Long oid;
 
   private String paymentId;
